feat: Extract mutations into separate server action folders
